+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add: initial data template files for jsonjsdb, static-make, and update-app
 - add: alternative db for testing
+- change: delivery format to just "format" in ui
 - fix: jsonjsdb with minimal data
 - fix: HTTP server compatibility with Windows network drives (SMB shares)
 - remove: md extension from LICENSE file